requirements.txt
--index-url https://download.pytorch.org/whl/cu118
torch==2.2.1+cu118
numpy<2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
wandb>=0.15.0
tqdm>=4.65.0
